Simplify Service Worker caching
restaurant.html is the same for all restaurants, and the information is
populated programmaticaly with JavaScript. Therefore, I only need to
cache restaurant.html once, instead of for every restaurant.

Note that this reduces caching of map tiles, but they're not fully
caching anyway. Static map images could be used for fully offline maps.